# Conflicts:
#	spring-backend/src/main/resources/application.properties

+++ spring-backend/src/main/java/backend/hobbiebackend/service/HobbyService.java
@@ -1,18 +1,17 @@
 package backend.hobbiebackend.service;
 
-import backend.hobbiebackend.model.entities.AppClient;
-import backend.hobbiebackend.model.entities.Hobby;
+import backend.hobbiebackend.entities.AppClient;
+import backend.hobbiebackend.entities.Hobby;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Set;
 
 public interface HobbyService {
-    Hobby findHobbieById(Long id);
+    Hobby findHobbieById(Integer id);
 
     void saveUpdatedHobby(Hobby hobby) throws Exception;
 
-    boolean deleteHobby(long id) throws Exception;
+    boolean deleteHobby(int id) throws Exception;
 
     Set<Hobby> findHobbyMatches(String username);
 
@@ -20,7 +19,7 @@ public interface HobbyService {
 
     boolean removeHobbyForClient(Hobby hobby, String username);
 
-    boolean isHobbySaved(Long hobbyId, String username);
+    boolean isHobbySaved(Integer hobbyId, String username);
 
     List<Hobby> findSavedHobbies(AppClient appClient);
